+++ src/main/java/de/maxhenkel/voicechat/voice/client/HUDIconLocation.java
@@ -1,0 +1,7 @@
+package de.maxhenkel.voicechat.voice.client;
+
+public enum HUDIconLocation {
+
+    LEFT, RIGHT, CENTER
+
+}